add proper health check
The current health check only indicates that the server is running
without checking that the DNS records are indeed up to date.

To make /healthz more useful, we additionally support ipv4 and ipv6
parameters. If any of them is provided, the DNS record in question
is fetched and the addresses are compared to what was specified in the
ipv4 or ipv6 parameter arguments. If the parameter arguments were empty,
the address is instead fetched via https://api.ipify.org.